--- Client-Raspi/src/main/org/kabieror/elwasys/raspiclient/devices/DeconzChangeType.java
@@ -1,5 +1,0 @@
-package org.kabieror.elwasys.raspiclient.devices;
-
-public enum DeconzChangeType {
-    changed
-}
