--- Client-Raspi/src/main/org/kabieror/elwasys/raspiclient/devices/IDeconzPowerMeasurementEventListener.java
@@ -1,5 +1,0 @@
-package org.kabieror.elwasys.raspiclient.devices;
-
-public interface IDeconzPowerMeasurementEventListener {
-    void onPowerMeasurementReceived(DeconzPowerMeasurementEvent event);
-}
