--- Client-Raspi/src/main/org/kabieror/elwasys/raspiclient/devices/IDeconzSessionClosedEventListener.java
@@ -1,5 +1,0 @@
-package org.kabieror.elwasys.raspiclient.devices;
-
-public interface IDeconzSessionClosedEventListener {
-    void onDeconzSessionClosed();
-}
