+++ Client-Raspi/src/main/org/kabieror/elwasys/raspiclient/devices/deconz/DeconzChangeType.java
@@ -1,0 +1,5 @@
+package org.kabieror.elwasys.raspiclient.devices.deconz;
+
+enum DeconzChangeType {
+    changed
+}
