+++ Client-Raspi/src/main/org/kabieror/elwasys/raspiclient/devices/deconz/IDeconzPowerMeasurementEventListener.java
@@ -1,0 +1,5 @@
+package org.kabieror.elwasys.raspiclient.devices.deconz;
+
+interface IDeconzPowerMeasurementEventListener {
+    void onPowerMeasurementReceived(DeconzPowerMeasurementEvent event);
+}
